mcp/backend/requirements.txt
# MCP Framework
mcp==1.13.1
click>=8.2.0
httpx>=0.27
pydantic>=2.0
pydantic-settings>=2.5.2
starlette>=0.28.0
uvicorn[standard]>=0.31.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0

# API & Utilities
requests>=2.32.5
python-dotenv>=1.0.0

# Visualization (for reports)
plotly>=5.0.0
